Conversion for win paths in URIs, wip again (now compiles on win).
git-svn-id: svn://falconpl.org/falcon/core/trunk@1312 78e5f83a-d9f7-4e86-b854-dad15b5eb37a

+++ engine/core_module/table.cpp
@@ -681,7 +681,7 @@ FALCON_FUNC  Table_insert ( ::Falcon::VMachine *vm )
       pos = page->length();
    else
    {
-      pos = i_pos->forceInteger();
+      pos = (uint32) i_pos->forceInteger();
 
       // insert a bottom?
       if( pos > page->length() )
